requirements.txt
pillow
pytesseract
ollama
pydantic>=2
python-dateutil
pytest
opencv-python
numpy
scipy
